@@ -2,7 +2,7 @@ package com.example.studentform_with_fragments;
 
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 
 public class MainActivity extends AppCompatActivity implements MainFragment.OnFragmentInteractionListener,
