@@ -5,8 +5,6 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -14,8 +12,6 @@ import android.view.View;
 
 import android.widget.ImageButton;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -27,7 +23,7 @@ import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
 
-    private ImageButton start_button;
+    private static final String TAG = "MainActivityTag";
 
     private RecyclerView mRecyclerView;
     private RecyclerView.Adapter mAdapter;
@@ -37,30 +33,45 @@ public class MainActivity extends AppCompatActivity {
     private List<Meeting> mMeetings = new ArrayList<>();
 
 
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Log.d("Test", "App started");
+        Log.d(TAG, "App started");
 
         mDatabase = FirebaseDatabase.getInstance().getReference();
 
-        start_button = findViewById(R.id.add_imageButton);
+        ImageButton start_button = findViewById(R.id.add_imageButton);
 
         mRecyclerView = findViewById(R.id.recyclerView);
         mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
-        getMeetings();
-//        String id = "21";
-//        Meeting m = new Meeting("Title1","Woodward",id,2020,12,10,11,13);
-//        Log.d("Test", "M:  " + m.toString());
+
+//        mMeetings.add(new Meeting("Title1","Woodward","1",2020,12,10,11,13));
+//        mMeetings.add(new Meeting("Title1","Woodward","2",2020,12,10,11,13));
 //
-//        mMeetings.add(m);
-//        mDatabase.child("meetings").setValue(id);
-//        mDatabase.child("meetings").child(id).setValue(m);
-//        Log.d("Test", "Inside " + mDatabase.child("meetings"));
+//        mDatabase.child("meetings").setValue(mMeetings);
 
 
+        Log.d(TAG, "onCreate: ");
+        Log.d(TAG, "mMeetings Pre" + mMeetings.size());
+        getMeetings(new FirebaseCallback() {
+            @Override
+            public void onCallback(List<Meeting> list) {
+                Log.d(TAG, "onCallback: "+list.toString());
+                mMeetings = list;
+                Log.d(TAG, "mMeetings Post" + mMeetings.size());
+
+                MainActivity.this.runOnUiThread(new Runnable() {
+                    public void run() {
+                        mAdapter = new MeetingAdapter(mMeetings, MainActivity.this);
+                        mRecyclerView.setAdapter(mAdapter);
+                        mAdapter.notifyDataSetChanged();
+                    }
+                });
+
+
+            }
+        });
 
         start_button.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -71,34 +82,27 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
-    public void getMeetings(){
+    public interface FirebaseCallback {
+        void onCallback(List<Meeting> list);
+    }
 
-        Log.d("Test", "getMeeting DB: " + mDatabase);
+    public void getMeetings(final FirebaseCallback firebaseCallback){
+
+        Log.d(TAG, "getMeeting DB: " + mDatabase);
         mDatabase.child("meetings").addListenerForSingleValueEvent(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
 
-                Log.d("Test", "getMeeting RV: " + mRecyclerView);
+                Log.d(TAG, "getMeeting RV: " + dataSnapshot.getChildrenCount());
                 for (DataSnapshot ds : dataSnapshot.getChildren()) {
-                    String key = ds.getKey();
-
-//                    Log.d("Test", "New " + ds.getValue(String.class));
-                    Log.d("Test", "onDataChange: "+ds.getValue().toString());
+                    mMeetings.add(ds.getValue(Meeting.class));
                 }
+                firebaseCallback.onCallback(mMeetings);
             }
 
             @Override
             public void onCancelled(@NonNull DatabaseError databaseError) {
-
-            }
-        });
-
-
-        MainActivity.this.runOnUiThread(new Runnable() {
-            public void run() {
-                mAdapter = new MeetingAdapter(mMeetings, MainActivity.this);
-                mRecyclerView.setAdapter(mAdapter);
-                mAdapter.notifyDataSetChanged();
+                Log.d(TAG, "onCancelled: " + databaseError.getMessage());
             }
         });
     }
