@@ -1,6 +1,5 @@
 package com.example.newsappxmlrss;
 
-import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -8,7 +7,6 @@ import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.net.Uri;
 import android.os.AsyncTask;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -19,9 +17,11 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.squareup.picasso.Picasso;
 
-import org.apache.commons.io.IOUtils;
 import org.xmlpull.v1.XmlPullParserException;
 
 import java.io.IOException;
@@ -200,7 +200,7 @@ public class MainActivity extends AppCompatActivity {
 
     }
     private void setArticle(Article current_article) {
-        Picasso.with(MainActivity.this).load(current_article.getUrlToImage()).into(imageView);
+        Picasso.get().load(current_article.getUrlToImage()).into(imageView);
         descrip.setText(current_article.getDescription());
         title.setText(current_article.getTitle());
         publishedAt.setText(current_article.getPublishedAt());
